Implement new SQLCipher based database and history
qTox will automatically import the old history on startup.

This new database code is much more robust.
It is very resilient and will not corrupt or disappear after a crash or
power failure, unlike the old code.
The on-disk database format is also much more compact now.

The database sync option in the advanced settings has been removed,
we know run many database operations asynchronously so performance
should not be a problem anymore, but we always ensure resiliency
in case of abrupt termination, so there is no tradeoff anymore.

+++ src/persistence/offlinemsgengine.h
@@ -39,7 +39,7 @@ public:
     static QMutex globalMutex;
 
     void dischargeReceipt(int receipt);
-    void registerReceipt(int receipt, int messageID, ChatMessage::Ptr msg, const QDateTime &timestamp = QDateTime::currentDateTime());
+    void registerReceipt(int receipt, int64_t messageID, ChatMessage::Ptr msg, const QDateTime &timestamp = QDateTime::currentDateTime());
 
 public slots:
     void deliverOfflineMsgs();
@@ -54,8 +54,8 @@ private:
 
     QMutex mutex;
     Friend* f;
-    QHash<int, int> receipts;
-    QMap<int, MsgPtr> undeliveredMsgs;
+    QHash<int, int64_t> receipts;
+    QMap<int64_t, MsgPtr> undeliveredMsgs;
 
     static const int offlineTimeout;
 };
